+++ koin-projects/koin-core/src/main/kotlin/org/koin/core/Koin.kt
@@ -47,8 +47,8 @@ class Koin {
      */
     @JvmOverloads
     inline fun <reified T> inject(
-            qualifier: Qualifier? = null,
-            noinline parameters: ParametersDefinition? = null
+        qualifier: Qualifier? = null,
+        noinline parameters: ParametersDefinition? = null
     ): Lazy<T> = rootScope.inject(qualifier, parameters)
 
     /**
@@ -61,8 +61,8 @@ class Koin {
      */
     @JvmOverloads
     inline fun <reified T> injectOrNull(
-            qualifier: Qualifier? = null,
-            noinline parameters: ParametersDefinition? = null
+        qualifier: Qualifier? = null,
+        noinline parameters: ParametersDefinition? = null
     ): Lazy<T?> = rootScope.injectOrNull(qualifier, parameters)
 
     /**
@@ -73,8 +73,8 @@ class Koin {
      */
     @JvmOverloads
     inline fun <reified T> get(
-            qualifier: Qualifier? = null,
-            noinline parameters: ParametersDefinition? = null
+        qualifier: Qualifier? = null,
+        noinline parameters: ParametersDefinition? = null
     ): T = rootScope.get(qualifier, parameters)
 
     /**
@@ -87,8 +87,8 @@ class Koin {
      */
     @JvmOverloads
     inline fun <reified T> getOrNull(
-            qualifier: Qualifier? = null,
-            noinline parameters: ParametersDefinition? = null
+        qualifier: Qualifier? = null,
+        noinline parameters: ParametersDefinition? = null
     ): T? = rootScope.getOrNull(qualifier, parameters)
 
     /**
@@ -101,11 +101,26 @@ class Koin {
      * @return instance of type T
      */
     fun <T> get(
-            clazz: KClass<*>,
-            qualifier: Qualifier?,
-            parameters: ParametersDefinition?
+        clazz: KClass<*>,
+        qualifier: Qualifier?,
+        parameters: ParametersDefinition?
     ): T = rootScope.get(clazz, qualifier, parameters)
 
+
+    /**
+     * Declare a component definition from the given instance
+     * This result of declaring a single definition of type T, returning the given instance
+     *
+     * @param instance
+     * @param qualifier
+     */
+    inline fun <reified T> declare(
+        instance: T,
+        qualifier: Qualifier? = null,
+        secondaryTypes: List<KClass<*>>? = null
+    ) {
+        rootScope.declare(instance, qualifier,secondaryTypes)
+    }
 
     /**
      * Get a all instance for given inferred class (in primary or secondary type)
@@ -120,7 +135,8 @@ class Koin {
      *
      * @return instance of type S
      */
-    inline fun <reified P, reified S> bind(noinline parameters: ParametersDefinition? = null): S = rootScope.bind<P, S>(parameters)
+    inline fun <reified P, reified S> bind(noinline parameters: ParametersDefinition? = null): S =
+        rootScope.bind<P, S>(parameters)
 
     /**
      * Get instance of primary type P and secondary type S
@@ -128,7 +144,11 @@ class Koin {
      *
      * @return instance of type S
      */
-    fun <S> bind(primaryType: KClass<*>, secondaryType: KClass<*>, parameters: ParametersDefinition? = null): S = rootScope.bind(primaryType, secondaryType, parameters)
+    fun <S> bind(
+        primaryType: KClass<*>,
+        secondaryType: KClass<*>,
+        parameters: ParametersDefinition? = null
+    ): S = rootScope.bind(primaryType, secondaryType, parameters)
 
     internal fun createEagerInstances() = rootScope.createEagerInstances()